pyproject.toml
[tool.poetry]
name = "habij"
version = "0.1.0"
description = ""
authors = ["thevahidal <thevahidal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I"]
